requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
bcrypt==4.2.1
bidict==0.23.1
boto3==1.35.94
botocore==1.35.94
branca==0.8.1
build==1.2.2.post1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clipboard==0.0.4
cryptography==44.0.0
deprecation==2.1.0
docutils==0.21.2
ecdsa==0.19.0
fastapi==0.115.6
folium==0.19.4
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
gotrue==2.11.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
jmespath==1.0.1
keyring==25.6.0
lazy_loader==0.4
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
mysql-connector-python==9.1.0
mysqlclient==2.2.6
nh3==0.2.20
numpy==2.2.1
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pipdeptree==2.16.2
pkginfo==1.12.0
platformdirs==4.3.6
plotly==5.24.1
postgrest==0.19.1
propcache==0.2.1
psutil==6.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.1
PyMySQL==1.1.1
pyperclip==1.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-jose==3.3.0
python-multipart==0.0.20
python-socketio==5.12.1
pytz==2024.2
PyYAML==6.0.2
readme_renderer==44.0
realtime==2.1.0
redis==5.2.1
reflex==0.6.8
reflex-chakra==0.6.2
reflex-hosting-cli==0.1.34
reflex-local-auth==0.2.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
setuptools==75.7.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
starlette-admin==0.14.1
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.0
tabulate==0.9.0
tenacity==9.0.0
tomlkit==0.13.2
twine==6.0.1
typer==0.15.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==13.1
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
xyzservices==2024.9.0
yarl==1.18.3
